+++ nl.cwi.dis.vr2gather/Runtime/VRTCommon/Players/PositionTracker.cs
@@ -1,12 +1,11 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Data;
 using System.IO;
 using Cwipc;
 using JetBrains.Annotations;
 using Unity.Profiling;
-using UnityEditor.EditorTools;
+// using UnityEditor.EditorTools;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using VRT.Core;