+++ nl.cwi.dis.vr2gather/Runtime/VRTUserPointClouds/Scripts/LiveTileSelector.cs
@@ -1,5 +1,6 @@
+using Cwipc;
 using UnityEngine;
-using VRT.Pilots.Common;
+using VRT.Core;
 
 namespace VRT.UserRepresentation.PointCloud
 {
@@ -11,6 +12,7 @@ namespace VRT.UserRepresentation.PointCloud
         //
         // Temporary variable (should be measured from internet connection): total available bitrate for this run.
         //
+        [Tooltip("If non-zero: don't measure the bitrate budget but use this number")]
         public double bitRatebudget = 1000000;
         //
         // For live: we precompute the bandwidth usage matrix based on the reported
@@ -25,6 +27,9 @@ namespace VRT.UserRepresentation.PointCloud
             nTiles = _tilingConfig.tiles.Length;
             nQualities = _tilingConfig.tiles[0].qualities.Length;
             Debug.Log($"{Name()}: nQualities={nQualities}, nTiles={nTiles}");
+#if VRT_WITH_STATS
+            Statistics.Output(Name(), $"nQualities={nQualities}, nTiles={nTiles}");
+#endif
             TileOrientation = new Vector3[nTiles];
             for (int ti = 0; ti < nTiles; ti++)
             {
@@ -40,6 +45,16 @@ namespace VRT.UserRepresentation.PointCloud
                 }
             }
         }
+
+        new public void Start()
+        {
+            base.Start();
+            var settings = VRTConfig.Instance.TileSelector;
+            if (settings.bitrateBudget != 0) {
+                bitRatebudget = settings.bitrateBudget;
+            }
+        }
+        
 
         protected override double[][] getBandwidthUsageMatrix(long currentFrameNumber)
         {
@@ -60,10 +75,19 @@ namespace VRT.UserRepresentation.PointCloud
         {
             // xxxjack currently returns camera viedw angle (as the name implies)
             // but maybe camera position is better. Or both.
+#if xxxjack_disabled
+            // This code was dependent on the TileSelector being part of the P_Player_Self.
+            // But now it is on the P_Player for the other players.
             PlayerControllerSelf player = gameObject.GetComponentInParent<PlayerControllerSelf>();
             Transform cameraTransform = player?.getCameraTransform();
+#else
+            Transform cameraTransform = Camera.main.transform;
+#endif
             if (cameraTransform == null)
+            {
                 Debug.LogError($"{Name()}: Camera not found");
+                return Vector3.forward;
+            }
             return cameraTransform.forward;
 
         }