+++ nl.cwi.dis.vr2gather/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- Preliminary WebRTC protocol support added (#166, #161, #102)
+- Tile quality selector added. Decisions are probably wrong. (#182)
+
 ## [1.0.4] - 2024-10-23
 
 - Bug fixes (#178, #189)